src/requirements.txt
argparse==1.4.0
torch==2.3.0
pillow==10.3.0
transformers==4.40.2
Flask==3.0.3